Added AI
AI is working, but its playing is not that great. I will continue to
work on the the evaluation functions.

Right now the evaluation is the summation of values from these four
functions:
1. Landing Height: The height where the piece is put (= the height of
the column + (the height of the piece / 2))
2. Rows eliminated: The number of rows eliminated.
3. Row Transitions: The total number of row transitions. A row
transition occurs when an empty cell is adjacent to a filled cell on the
same row and vice versa.
4. Column Transitions: The total number of column transitions. A column
transition occurs when an empty cell is adjacent to a filled cell on the
same column and vice versa.